Fixed most of the implementation comments
